+++ src/mmgpy/ui/__init__.py
@@ -18,7 +18,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from mmgpy import Mesh
+    from mmgpy._mesh import Mesh
     from mmgpy.ui.app import MmgpyApp